server/requirements.txt
flask
flask-cors
requests
python-dotenv
PyPDF2
sentence-transformers
faiss-cpu
numpy
